+++ nirum/func.py
@@ -1,20 +1,6 @@
-import re
+from six.moves import urllib
 
-from six.moves import urllib, reduce
-
-__all__ = 'IMPORT_RE', 'import_string', 'url_endswith_slash'
-IMPORT_RE = re.compile(
-    r'''^
-        (?P<modname> (?!\d) [\w]+
-                     (?: \. (?!\d)[\w]+ )*
-        )
-        :
-        (?P<clsexp> (?P<clsname> (?!\d) \w+ )
-                    (?: \(.*\) )?
-        )
-    $''',
-    re.X
-)
+__all__ = 'url_endswith_slash'
 
 
 def url_endswith_slash(url):
@@ -24,28 +10,3 @@ def url_endswith_slash(url):
     if not path.endswith('/'):
         path += '/'
     return urllib.parse.urlunsplit((scheme, netloc, path, '', ''))
-
-
-def import_string(imp):
-    """Don't use this.
-
-    .. deprecated:: 0.6.0
-       It will be completely obsolete at version 0.7.0.
-
-    """
-    m = IMPORT_RE.match(imp)
-    if not m:
-        raise ValueError(
-            "malformed expression: {}, have to be x.y:z(...)".format(imp)
-        )
-    module_name = m.group('modname')
-    import_root_mod = __import__(module_name)
-    # it is used in `eval()`
-    import_mod = reduce(getattr, module_name.split('.')[1:], import_root_mod) # noqa
-    class_expression = m.group('clsexp')
-    try:
-        v = eval(class_expression, import_mod.__dict__, {})
-    except AttributeError:
-        raise ValueError("Can't import {}".format(imp))
-    else:
-        return v